Turtle

Description

proper module documentation here

The turtle type(s)

type Program = Turtle

data Turtle

Constructors

Turtle

Fields

position :: Position

direction :: Angle

pen :: Bool

col :: Color

Primitive operations

forward :: Program -> Double -> Program

Turtle take a step forward.

right :: Program -> Double -> Program

Turn the turtle right with given angle.

penup :: Program -> Program

Perform commands one after another. (>*>) :: Program -> (Program -> a -> Program) -> a -> Program (turtle1 >*> turtle2) x = turtle1 x >>= turtle2

Penup: stop drawing turtle

pendown :: Program -> Program

Pendown: start drawing turtle

color :: Program -> Color -> Program

Changes the color of the turtle's pen.

die :: Program -> Program

Kills the turtle.

idle :: Program -> Program

A program that does nothing.

limited :: Program -> Int -> Program

Makes the turtle stop what it is doing after a specified period of time.

lifespan :: Program -> Int -> Program

Kills the turtle after a specified period of time.

backward :: Program -> Double -> Program

Turtle take a step back.

left :: Program -> Double -> Program

Turn the turtle left with given angle.

times :: Program -> Int -> Program

Repeats a turtle program a certain number of times.

forever :: Program -> Program

Repeats a program forever.

Derived operations

Run functions